+++ components/views/ScanConnect/ScanConnectDesktop.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex, Spinner, Stack, Text, Tooltip } from '@chakra-ui/react'
+import { Box, Flex, HStack, Spinner, Stack, Text } from '@chakra-ui/react'
 import { Wallet } from '../../../data/wallets'
 import QRCode from '../../QRCode'
 import CopyButton from '../../CopyButton'
@@ -7,7 +7,6 @@ import { useWcUri } from '../../../hooks/useWcUri'
 import { useWalletHistory } from '../../../hooks/useWalletHistory'
 import { handleCancel } from '../../../helpers/window'
 import { ViewContainer } from '../../layout/ViewContainer'
-import { ReactNode, useEffect, useState } from 'react'
 
 interface ScanConnectDesktopProps {
   wallet: Wallet
@@ -23,31 +22,6 @@ export default function ScanConnectDesktop({
     setLastUsed(wallet)
     handleCancel()
   })
-
-  const [isClipboardAllowed, setIsClipboardAllowed] = useState(true)
-
-  useEffect(() => {
-    navigator.permissions
-      .query({ name: 'clipboard-write' as PermissionName })
-      .then(result => {
-        setIsClipboardAllowed(result?.state === 'granted')
-      })
-      .catch(() => {
-        setIsClipboardAllowed(false)
-      })
-  }, [])
-
-  const wrapTooltip = (node: ReactNode) => {
-    if (isClipboardAllowed) {
-      return node
-    }
-
-    return (
-      <Tooltip label={isClipboardAllowed ? '' : 'Clipboard access is blocked'}>
-        {node}
-      </Tooltip>
-    )
-  }
 
   if (connecting) {
     return (
@@ -67,15 +41,10 @@ export default function ScanConnectDesktop({
       spacing={2}
       justifyContent="space-evenly"
     >
-      <Flex justifyContent="space-between" width="100%" alignItems="center">
+      <HStack width="full">
         <Text textStyle="body2">Scan in the {wallet.name} app to connect</Text>
-        {wrapTooltip(
-          <CopyButton
-            text={uri}
-            disabled={!uri || isLoading || !isClipboardAllowed}
-          />,
-        )}
-      </Flex>
+        <CopyButton text={uri} isDisabled={!uri || isLoading} ml="auto" />
+      </HStack>
 
       <Box padding={3} borderRadius="0.75rem" borderWidth="1px" bg="white">
         {uri && (
